FEAT: AwardsItemDot 컴포넌트 추가
